fail-open: Add new fail safe mode 'standalone-loose'
Currently, the standalone fail safe mode will add a NORMAL flow when the
controller is disconnected for a predefined period, and will remove the
NORMAL flow right away once the controller is reconnected.  Thus, user
may observe short disconnectivity from the time that the NORMAL flow is
removed till the controller write back the flows.

To avoid the disconnectivity during the timespan, this patch introduces
a new fail safe mode, standalone-loose, which behaves similarly as the
standalone mode, but it keeps the NORMAL flow when the controller is
reconnected.  It let the controller to decide whether to remote the
NORMAL flow later on.

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>