ofp-actions: Add support for conntrack timeout
This patch adds support for specifying a timeout policy for a
connection in connection tracking system. The timeout policy
is attached to a connection when the connection is first committed
to conntrack.

Currently, this feature is available in Linux kernel datapath.
Here is an example of configuring a timeout policy in Linux kernel
with nfct, and sample flows of attaching the timeout policy with a
connection.

    $ nfct add timeout ovs_timeout_1 inet icmp timeout 10

    table=0,priority=1,action=drop
    table=0,arp,action=normal
    table=0,in_port=1,icmp,action=ct(commit,timeout=1),2
    table=0,in_port=2,icmp,action=ct(table=1)
    table=1,in_port=2,icmp,ct_state=+trk+est,action=1

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>