system-traffic: Add conntrack timeout system traffic test
This patch adds a system traffic test to verify the conntrack
timeout feature.  The test uses nfct utility to configure
the customized ICMP and UDP timeout to a short period.  It then
injects ICMP and UDP traffic to conntrack, and checks if the
corresponding conntrack entry expires after the predefined timeout.

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>