openvswitch: Allow attaching helper in later commit
This patch allows to attach conntrack helper to a confirmed conntrack
entry.  Currently, we can only attach alg helper to a conntrack entry
when it is in the unconfirmed state.  This patch enables an use case
that we can firstly commit a conntrack entry after it passed some
initial checks.  After that the processing pipeline will further
check if the connection belongs to a particular application, and attach
alg helper to the connection in the later stage.

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>