netdev_afxdp: Detects combined channels and aborts wrong config
This patches detects the number of combined channels on a AF_XDP port
via using ethtool interface.  If the number of combined channels
is different from the n_rxq, netdev_afxdp will not be able to get
all the packets from that port.  Thus, abort the port set up when
the case is detected.

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>